Add animated scroll to validation errors
I find it confusing to click the Publish/Update button and not see a result. This happens when validation errors occur - the validation fails and error messages are outputted, but no visual cues are given to the user if the error messages are below the fold.
This will scroll to the topmost error message on Publish/Update.