Add rule for shortcodes
Allow field groups to be placed on posts that have or do not have,
specified shortcodes.
Added a new class for the select2 to use: "select2-custom", which allows
the user to define their own value in case it doesn't exist in the list.
See the input created for the shortcode on usage.