Add action to append options to all fields
I am wanting to add some global options to all of my fields, but to my dismay there were no actions inside of the field_html renderer that would allow you to do this. Currently the only way to have custom field_html is to extend a field in a subclass, but that doesn't work when you want to add global options to all fields.